refactor: query builder
