+++ src/core/features/notes/controller/NotesController.ts
@@ -1,6 +1,7 @@
 /* eslint-disable camelcase */
 import { v4 as uuid4 } from 'uuid';
-import { qb, wrapDB } from '@utils/db/wrapDB';
+import { qb } from '@utils/db/query-builder';
+import { wrapDB } from '@utils/db/wrapDB';
 
 import { SQLiteDatabase } from '../../../storage/database/SQLiteDatabase/SQLiteDatabase';
