+++ src/utils/db/query-builder.ts
@@ -1,0 +1,4 @@
+import { SQLCompiler } from 'nano-queries/compilers/SQLCompiler';
+import { ConfigurableSQLBuilder } from 'nano-queries/sql/ConfigurableSQLBuilder';
+
+export const qb = new ConfigurableSQLBuilder(new SQLCompiler());
